requirements.txt
Flask
flask-cors
pymongo
PyJWT
python-dotenv
pandas
numpy
openpyxl
gunicorn  
matplotlib
PyQt5